repoboard/whatsapp-bot/requirements.txt
# Option 1: WhatsApp Web (Unofficial)
whatsapp-web.py==0.1.0

# Option 2: Twilio (Official, Recommended)
flask==3.0.0
twilio==8.10.0
requests==2.31.0

